+++ src/OpaDotNet.Extensions.AspNetCore/IOpaImportsAbiFactory.cs
@@ -5,4 +5,6 @@ namespace OpaDotNet.Extensions.AspNetCore;
 public interface IOpaImportsAbiFactory
 {
     Func<IOpaImportsAbi> ImportsAbi { get; }
+
+    Func<Stream>? Capabilities { get; }
 }