+++ src/OpaDotNet.Extensions.AspNetCore/OpaImportsAbiFactory.cs
@@ -6,6 +6,8 @@ internal class OpaImportsAbiFactory : IOpaImportsAbiFactory
 {
     public Func<IOpaImportsAbi> ImportsAbi { get; }
 
+    public Func<Stream>? Capabilities { get; }
+
     internal OpaImportsAbiFactory()
     {
         ImportsAbi = () => new CoreImportsAbi();
@@ -14,6 +16,16 @@ internal class OpaImportsAbiFactory : IOpaImportsAbiFactory
     public OpaImportsAbiFactory(Func<IOpaImportsAbi> importsAbi)
     {
         ArgumentNullException.ThrowIfNull(importsAbi);
+
         ImportsAbi = importsAbi;
+    }
+
+    public OpaImportsAbiFactory(Func<IOpaImportsAbi> importsAbi, Func<Stream> capabilities)
+    {
+        ArgumentNullException.ThrowIfNull(importsAbi);
+        ArgumentNullException.ThrowIfNull(capabilities);
+
+        ImportsAbi = importsAbi;
+        Capabilities = capabilities;
     }
 }